Moved Hero comp
