+++ src/components/Hero.tsx
