+++ src/pages/product-detail.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
+import Hero from '~/components/Hero';
 import ProductSection from '~/components/product-detail/ProductSection';
-import Hero from '~/components/product-listing/Hero';
 import Container from '~/layouts/Container';
 
 const ProductDetail = () => {
