+++ src/pages/product-listing.tsx
@@ -1,6 +1,6 @@
+import Hero from '~/components/Hero';
 import ProductList from '~/components/ProductList';
 import Filter from '~/components/product-listing/Filter';
-import Hero from '~/components/product-listing/Hero';
 import Pagination from '~/components/product-listing/Pagination';
 
 const ProductListing = () => {
